Render Split - building lib-tempo
