+++ src/lib-tempo/CMakeLists.txt
@@ -8,6 +8,7 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/Types.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentGridAi.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentHealth.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/ComponentRender.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreation.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreationClient.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreationServer.hpp"
@@ -18,7 +19,6 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/PlayerLocal.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/PlayerRemote.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/PlayerRemoteServer.hpp"
-	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/Render.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/TileMask.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/Transform.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/network/base.hpp"
@@ -27,6 +27,7 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/network/server.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemGridAi.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemHealth.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/SystemRender.hpp"
 )
 
 set(TEMPO_LIB_TEMPO_SOURCE
@@ -36,6 +37,7 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/structures.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/time.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentHealth.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/entity/ComponentRender.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreation.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreationClient.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreationServer.cpp"
@@ -46,7 +48,6 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/PlayerLocal.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/PlayerRemote.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/PlayerRemoteServer.cpp"
-	"${CMAKE_CURRENT_LIST_DIR}/src/entity/Render.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/TileMask.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/Transform.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/network/base.cpp"
@@ -55,6 +56,7 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/network/server.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemGridAi.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemHealth.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/entity/SystemRender.cpp"
 )
 
 add_library(lib_tempo STATIC
