+++ src/lib-tempo/include/tempo/entity/ComponentRender.hpp
@@ -1,5 +1,5 @@
-#ifndef TEMPO_ENTITY_RENDER_HPP
-#define TEMPO_ENTITY_RENDER_HPP
+#ifndef TEMPO_COMPONENT_RENDER_HPP
+#define TEMPO_COMPONENT_RENDER_HPP
 
 #include <OgreRoot.h>
 #include <OgreSceneNode.h>
@@ -34,20 +34,6 @@ namespace tempo{
 		// Returns:
 		//          void
 		void AddHealthBar();
-	};
-
-	/////////////////////////////////////////////////////////////////////
-	/// \brief System that manages rendering entities, as well as managing
-	/// the Ogre scene
-	/////////////////////////////////////////////////////////////////////
-	struct SystemRender : anax::System<anax::Requires<ComponentTransform, ComponentRender>> {
-		Application&        app;
-		Ogre::SceneManager* scene;
-
-		SystemRender(Application& app);
-		~SystemRender();
-
-		void render(float dt);
 	};
 }
