+++ src/lib-tempo/include/tempo/entity/EntityCreationClient.hpp
@@ -8,7 +8,7 @@
 #include <tempo/component/ComponentGridAi.hpp>
 #include <tempo/component/ComponentHealth.hpp>
 #include <tempo/entity/EntityCreation.hpp>
-#include <tempo/entity/Render.hpp>
+#include <tempo/entity/ComponentRender.hpp>
 #include <tempo/entity/ID.hpp>
 #include <tempo/entity/LevelManager.hpp>
 #include <tempo/entity/PlayerLocal.hpp>
